Refactored Navigation, added shared props interface, refactored imports
