+++ src/common/models/index.ts
@@ -1,0 +1,2 @@
+export * from './app-pages.enum';
+export * from './navigation-base-props';
