+++ src/common/models/navigation-base-props.ts
@@ -1,0 +1,6 @@
+import { AppPages } from './app-pages.enum';
+
+export interface NavigationBaseProps {
+    navigate: (nextPage: AppPages) => void;
+    currentPage: AppPages;
+}
