+++ src/pages/Order/Order.tsx
@@ -1,3 +1,3 @@
 import React from 'react';
 
-export const Order = () => <p>Order Page</p>
+export const Order = () => <p>Map Page</p>
